+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for SPOC
 - PDU add metadata block
 - Support for AISTECHSAT-2 custom protocol
+- Lucky-7 image receiver
 
 ### Changed
 - Replaced boost::bind() by C++ lambdas